- Implemented delete functionality on admin page.

+++ Cadet-Uniform-IMS/Pages/Admin/BrowseUniform_Types.cshtml.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Authorization;
 
 namespace Cadet_Uniform_IMS.Pages.Admin
 {
-    [Authorize(Roles ="Admin")]
+    [Authorize(Roles = "Admin")]
     public class BrowseUniformModel : PageModel
     {
         private readonly Cadet_Uniform_IMS.Data.IMS_Context _context;
@@ -20,7 +20,7 @@ namespace Cadet_Uniform_IMS.Pages.Admin
             _context = context;
         }
 
-        public IList<Uniform> Uniform { get;set; } = default!;
+        public IList<Uniform> Uniform { get; set; } = default!;
 
         public IList<UniformType> Types { get; set; } = default!;
 
@@ -35,46 +35,104 @@ namespace Cadet_Uniform_IMS.Pages.Admin
 
         }
 
-        public async Task<IActionResult> OnDeleteType(int id)
+        public async Task<IActionResult> OnPostDeleteTypeAsync(int id)
         {
-            var uniforms = await _context.Uniform.Where(i => i.TypeID == id).ToListAsync();
+            var uniforms = await _context.Uniform
+                .Where(u => u.TypeID == id)
+                .ToListAsync();
+
             if (uniforms.Any())
             {
                 foreach (var uniform in uniforms)
                 {
-                    var stocks = await _context.Stock.Where(i => i.UniformID == uniform.UniformID ).ToListAsync();
+                    var stocks = await _context.Stock
+                        .Where(s => s.UniformID == uniform.UniformID)
+                        .ToListAsync();
+
                     if (stocks.Any())
                     {
-                        foreach (var stock in stocks) {
-                            var StockSize = await _context.StockSize.Where(i => i.StockID == stock.StockID).ToListAsync();
-                            if (!StockSize.Any())
-                            {
-                                foreach (var size in StockSize)
-                                {
-                                    _context.StockSize.Remove(size);
-                                }
-                            }
-                            _context.Stock.Remove(stock);
+                        foreach (var stock in stocks)
+                        {
+                            var stockSizes = await _context.StockSize
+                                .Where(ss => ss.StockID == stock.StockID)
+                                .ToListAsync();
+
+                            _context.StockSize.RemoveRange(stockSizes);
+                            await _context.SaveChangesAsync();
                         }
+
+                        _context.Stock.RemoveRange(stocks);
                     }
+
                     _context.Uniform.Remove(uniform);
                 }
             }
 
-            var attributes = await _context.SizeAttribute.Where(i => i.TypeID == id).ToListAsync();
+            var attributes = await _context.SizeAttribute
+                .Where(i => i.TypeID == id)
+                .ToListAsync();
+
             if (attributes.Any())
             {
                 foreach (var attribute in attributes)
                 {
-                    _context.SizeAttribute.Remove(attribute);
+                    var stockSizes = await _context.StockSize
+                        .Where(ss => ss.AttributeID == attribute.AttributeID)
+                        .ToListAsync();
+
+                    _context.StockSize.RemoveRange(stockSizes);
                 }
+
+                _context.SizeAttribute.RemoveRange(attributes);
             }
 
             var type = await _context.UniformType.FindAsync(id);
-            _context.UniformType.Remove(type);
+            if (type != null)
+            {
+                _context.UniformType.Remove(type);
+            }
+
+            await _context.SaveChangesAsync();
             await OnGet();
             return Page();
         }
 
+        public async Task<IActionResult> OnPostDeleteUniformAsync(int id)
+        {
+            var uniforms = await _context.Uniform
+                .Where(u => u.UniformID == id)
+                .ToListAsync();
+
+            if (uniforms.Any())
+            {
+                foreach (var uniform in uniforms)
+                {
+                    var stocks = await _context.Stock
+                        .Where(s => s.UniformID == uniform.UniformID)
+                        .ToListAsync();
+
+                    if (stocks.Any())
+                    {
+                        foreach (var stock in stocks)
+                        {
+                            var stockSizes = await _context.StockSize
+                                .Where(ss => ss.StockID == stock.StockID)
+                                .ToListAsync();
+
+                            _context.StockSize.RemoveRange(stockSizes);
+                            await _context.SaveChangesAsync();
+                        }
+
+                        _context.Stock.RemoveRange(stocks);
+                    }
+
+                    _context.Uniform.Remove(uniform);
+                }
+            }
+
+            await _context.SaveChangesAsync();
+            await OnGet();
+            return Page();
+        }
     }
 }
